requirements.txt
asgiref==3.8.1
contourpy==1.2.0
crispy-bootstrap5==2024.2
cycler==0.12.1
dj-database-url==2.1.0
Django==5.0.3
django-appconf==1.0.6
django-crispy-forms==2.1
django-select2==8.1.2
fonttools==4.50.0
gunicorn==21.2.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
psycopg2-binary==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
whitenoise==6.6.0
